Attempt at API Template
Keep in mind I have no idea what I'm looking at half of the time. If you can't make sense of what i threw together:

I used @xXEthyleneXx's api template and used it to eventually return a single user based on the login that has not been built yet. Basically this has no use as I don't have access to a DB and don't know the variable names, but it's halfway ready for one to come. The api will eventually return name, roll, image, blah and fill in the user[] in the javascript request file. Hopefully from there we can fill in the php file with the returned info from the DB.

As far as right now I don't even think I have it organized very well, and cant test it yet. I'll work on cloning this rep to vinyl so I can actually put it on a server to test php and links and stuff. I've realized that now that I have functions in, I need a functioning server instead of making my google chrome open up some blank files on my computer. We'll see.

Samuel Shelley
27 Jan 2023