Added Organization and Comments
Just tried to understand exactly whats going on. I think I've got the database figured out, but can't figure out what's in the array that it's returning, or how to access that array to then place it on the web page. Anyway, I'm running it on the 172...7003 and it seems to be going good

Samuel Shelley
28 January 2023